this removes some duplication in the "header button as survey" scheme
